refactor(framework): run framework tests through Vitest
AI-934. The 659-line smoke-framework.ts script becomes Vitest files: six
scorer tests next to the EVAL.ts they cover, four harness tests next to the
harness module they cover, shared plumbing in harness/scorer-test-kit.ts.
Adds a CI job so these actually run on PRs, which is the point of the ticket.

Colocation lets each scorer test import its scorer directly, so the old
dynamic loadScorer(path) is gone and the scorer/test contract is typechecked.

Two things the old script hid:

- Every failure was invisible. It stubbed console.error globally to silence
  expected-fail scorer noise, then reported failures through console.error, so
  a failing run printed nothing and exited 1. Vitest reports failures itself,
  so the stub moves to an onConsoleLog filter that cannot swallow a result.
- The frontend build test was failing for real (AI-975). Fixed here because a
  CI job that ships red is worse than no CI job: project-runner hardcoded
  <repo>/node_modules/{vite,vitest}, which pnpm's isolated layout never
  creates, and the frontend toolchain was declared in apps/framework where a
  scored workspace under results/ cannot resolve it. Bins now resolve via each
  package's exported package.json rather than a deep subpath that exports may
  stop publishing, and the workspace-facing toolchain moves to the root
  manifest, matching the documented contract that scoring runs repo-root
  vite/vitest so the sandbox need not carry it. Nothing in apps/framework
  imported those packages; they appeared only in generated workspace config
  and fixture strings. build-frontend-001-todos-app has never produced a
  result row, which is why nobody noticed.

The suite is hermetic: investigate-security-001's scorer calls a real OpenAI
judge, so that test mocks judge() alone and keeps every other core export
real. Verified green with OPENAI_API_KEY/ANTHROPIC_API_KEY unset.
