+++ apps/framework/package.json
@@ -9,7 +9,7 @@
     "eval:dry": "node --env-file=../../.env --import tsx/esm harness/run-eval.ts --dry",
     "eval:smoke": "node --env-file=../../.env --import tsx/esm harness/run-eval.ts --smoke",
     "typecheck": "tsc --noEmit",
-    "test:framework": "node --env-file-if-exists=../../.env --import tsx/esm scripts/smoke-framework.ts",
+    "test:framework": "vitest run",
     "export-results": "node --import tsx/esm scripts/export-results.ts",
     "demo:mcp": "node --env-file=../../.env --import tsx/esm scripts/mcp-demo.ts",
     "demo:executor": "node --env-file=../../.env --import tsx/esm scripts/executor-demo.ts"
@@ -22,16 +22,7 @@
     "@electric-sql/pglite": "catalog:",
     "@supabase-evals/core": "workspace:*",
     "@supabase-evals/sandbox": "workspace:*",
-    "@supabase/supabase-js": "catalog:",
-    "@testing-library/jest-dom": "^6.9.1",
-    "@testing-library/react": "^16.3.2",
-    "@vitejs/plugin-react": "catalog:",
     "ai": "catalog:",
-    "happy-dom": "^20.9.0",
-    "@supabase/lite": "catalog:",
-    "react": "^19.2.5",
-    "react-dom": "^19.2.5",
-    "vite": "catalog:",
     "vitest": "catalog:"
   },
   "devDependencies": {
