+++ apps/framework/tsconfig.json
@@ -5,6 +5,7 @@
     "harness",
     "shims",
     "scripts",
-    "../../evals/**/EVAL.ts"
+    "../../evals/**/EVAL.ts",
+    "../../evals/*/EVAL.test.ts"
   ]
 }
