+++ package.json
@@ -30,7 +30,17 @@
     "@supabase-evals/sandbox": "workspace:*",
     "@types/common-tags": "^1.8.4",
     "@types/node": "catalog:",
-    "typescript": "catalog:"
+    "typescript": "catalog:",
+    "@supabase/lite": "catalog:",
+    "@supabase/supabase-js": "catalog:",
+    "@testing-library/jest-dom": "^6.9.1",
+    "@testing-library/react": "^16.3.2",
+    "@vitejs/plugin-react": "catalog:",
+    "happy-dom": "^20.9.0",
+    "react": "^19.2.5",
+    "react-dom": "^19.2.5",
+    "vite": "catalog:",
+    "vitest": "catalog:"
   },
   "engines": {
     "node": ">=22",
